rft: 执行脚本
