+++ src/MaaWpfGui/Views/UserControl/Settings/ConnectSettingsUserControl.xaml
@@ -26,10 +26,6 @@
             <RowDefinition Height="*" />
             <RowDefinition Height="Auto" />
             <RowDefinition Height="*" />
-            <RowDefinition Height="*" />
-            <RowDefinition Height="*" />
-            <RowDefinition Height="*" />
-            <RowDefinition Height="*" />
         </Grid.RowDefinitions>
         <StackPanel
             Grid.Row="0"
@@ -252,49 +248,8 @@
             </StackPanel>
         </StackPanel>
 
-        <hc:TextBox
-            Grid.Row="7"
-            Width="300"
-            Height="30"
-            Margin="10"
-            hc:InfoElement.Placeholder="Example: &quot;C:\1.cmd&quot; -minimized"
-            hc:TitleElement.Title="{DynamicResource StartsWithScript}"
-            hc:TitleElement.TitlePlacement="Left"
-            AllowDrop="True"
-            Drop="StartsWithScript_Drop"
-            PreviewDragOver="TextBox_PreviewDragOver"
-            Text="{Binding StartsWithScript}" />
-        <hc:TextBox
-            Grid.Row="8"
-            Width="300"
-            Height="30"
-            Margin="10"
-            hc:InfoElement.Placeholder="Example: &quot;C:\1.cmd&quot; -noWindow"
-            hc:TitleElement.Title="{DynamicResource EndsWithScript}"
-            hc:TitleElement.TitlePlacement="Left"
-            AllowDrop="True"
-            Drop="EndsWithScript_Drop"
-            PreviewDragOver="TextBox_PreviewDragOver"
-            Text="{Binding EndsWithScript}" />
-        <WrapPanel
-            Grid.Row="9"
-            HorizontalAlignment="Center"
-            VerticalAlignment="Center">
-            <CheckBox
-                Margin="10"
-                HorizontalAlignment="Center"
-                VerticalAlignment="Center"
-                Content="{DynamicResource CopilotWithScript}"
-                IsChecked="{Binding CopilotWithScript}" />
-            <CheckBox
-                Margin="10"
-                HorizontalAlignment="Center"
-                VerticalAlignment="Center"
-                Content="{DynamicResource ManualStopWithScript}"
-                IsChecked="{Binding ManualStopWithScript}" />
-        </WrapPanel>
         <StackPanel
-            Grid.Row="10"
+            Grid.Row="7"
             HorizontalAlignment="Center"
             VerticalAlignment="Center"
             Orientation="Horizontal">
