+++ src/MaaWpfGui/Views/UserControl/Settings/GameSettingsUserControl.xaml
@@ -7,6 +7,7 @@
     xmlns:controls="clr-namespace:MaaWpfGui.Styles.Controls"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:dd="urn:gong-wpf-dragdrop"
+    xmlns:hc="https://handyorg.github.io/handycontrol"
     xmlns:helper="clr-namespace:MaaWpfGui.Helper"
     xmlns:local="clr-namespace:MaaWpfGui"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
@@ -71,5 +72,41 @@
             Margin="20,10"
             Content="{DynamicResource AutoRestartOption}"
             IsChecked="{Binding AutoRestartOnDrop}" />
+        <hc:TextBox
+            Grid.Row="7"
+            Width="300"
+            Height="30"
+            Margin="10"
+            hc:InfoElement.Placeholder="Example: &quot;C:\1.cmd&quot; -minimized"
+            hc:TitleElement.Title="{DynamicResource StartsWithScript}"
+            hc:TitleElement.TitlePlacement="Left"
+            AllowDrop="True"
+            Drop="StartsWithScript_Drop"
+            PreviewDragOver="TextBox_PreviewDragOver"
+            Text="{Binding StartsWithScript}" />
+        <hc:TextBox
+            Grid.Row="8"
+            Width="300"
+            Height="30"
+            Margin="10"
+            hc:InfoElement.Placeholder="Example: &quot;C:\1.cmd&quot; -noWindow"
+            hc:TitleElement.Title="{DynamicResource EndsWithScript}"
+            hc:TitleElement.TitlePlacement="Left"
+            AllowDrop="True"
+            Drop="EndsWithScript_Drop"
+            PreviewDragOver="TextBox_PreviewDragOver"
+            Text="{Binding EndsWithScript}" />
+        <CheckBox
+            Margin="10"
+            HorizontalAlignment="Center"
+            VerticalAlignment="Center"
+            Content="{DynamicResource CopilotWithScript}"
+            IsChecked="{Binding CopilotWithScript}" />
+        <CheckBox
+            Margin="10"
+            HorizontalAlignment="Center"
+            VerticalAlignment="Center"
+            Content="{DynamicResource ManualStopWithScript}"
+            IsChecked="{Binding ManualStopWithScript}" />
     </StackPanel>
 </UserControl>
