+++ src/MaaWpfGui/Views/UserControl/Settings/GameSettingsUserControl.xaml.cs
@@ -11,6 +11,9 @@
 // but WITHOUT ANY WARRANTY
 // </copyright>
 
+using System.Windows;
+using MaaWpfGui.ViewModels.UI;
+
 namespace MaaWpfGui.Views.UserControl.Settings
 {
     /// <summary>
@@ -24,6 +27,35 @@ namespace MaaWpfGui.Views.UserControl.Settings
         public GameSettingsUserControl()
         {
             InitializeComponent();
+        }
+
+        private void StartsWithScript_Drop(object sender, DragEventArgs e)
+        {
+            if (!e.Data.GetDataPresent(DataFormats.FileDrop))
+            {
+                return;
+            }
+
+            // Note that you can have more than one file.
+            var files = (string[])e.Data.GetData(DataFormats.FileDrop);
+            SettingsViewModel.GameSettings.StartsWithScript = files?[0] ?? string.Empty;
+        }
+
+        private void EndsWithScript_Drop(object sender, DragEventArgs e)
+        {
+            if (!e.Data.GetDataPresent(DataFormats.FileDrop))
+            {
+                return;
+            }
+
+            // Note that you can have more than one file.
+            var files = (string[])e.Data.GetData(DataFormats.FileDrop);
+            SettingsViewModel.GameSettings.EndsWithScript = files?[0] ?? string.Empty;
+        }
+
+        private void TextBox_PreviewDragOver(object sender, DragEventArgs e)
+        {
+            e.Handled = true;
         }
     }
 }
